Refactored the Controllers

+++ Backend/Backend/Services/EpisodeService.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Http.Connections;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Binders;
 using Microsoft.EntityFrameworkCore;
 using NAudio.Wave;
+using System;
 using System.ComponentModel;
 using System.Globalization;
 using System.Net.NetworkInformation;
@@ -83,11 +84,28 @@ namespace Backend.Services
 
 
         public async Task<bool> DeleteEpisode(Episode episode, DeleteEpisodeRequest deleteEpisodeRequest)
-        {  
-           await _fileService.DeleteFile(episode.AudioFileId.ToString()!);
-           _db.Remove(episode);
-           await _db.SaveChangesAsync();
-           return true;
+        {
+            Guid? guid = Guid.Parse(deleteEpisodeRequest.EpisodeId);
+            Files? file = await _db.File!.FirstOrDefaultAsync(u => u.FileId == episode.AudioFileId);
+
+            if (file != null)
+            {
+               var deleted = _fileService.DeleteFile(file.Path!);
+               if(deleted == true)
+                {
+                _db.Remove(file);
+
+                }
+               else { throw new Exception("Failed to Delete File"); }
+                _db.Remove(episode);
+                await _db.SaveChangesAsync();
+                return true;
+            }
+            else
+            {
+                throw new Exception("File Not Found");
+            }
+
 
         }
 
@@ -110,11 +128,18 @@ namespace Backend.Services
 
                         }
 
-                        bool? editted = await  _fileService.EditFile(episode.AudioFile!, editEpisodeRequest.AudioFile);
-                        if(editted == false)
+                        bool? editted =  _fileService.EditFile(episode.AudioFile!, editEpisodeRequest.AudioFile);
+                        if(editted == true)
+                        {
+                            episode.AudioFile!.Name = editEpisodeRequest.AudioFile.FileName;
+                            episode.AudioFile!.MimeType = editEpisodeRequest.AudioFile.ContentType;
+                            
+                           
+                        }
+                        else
                         {
                             throw new Exception("Failed to Edit the file");
-                           
+
                         }
